pyproject.toml
[project]
name = "skilltap-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "fastapi>=0.135.1",
    "jinja2>=3.1.6",
    "motor>=3.7.1",
    "pydantic-settings>=2.13.1",
    "pymongo>=4.16.0",
    "uvicorn>=0.42.0",
]
